# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "AIModel": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "recommended": {
            "default": false,
            "title": "Recommended",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ],
        "title": "AIModel",
        "type": "object"
      },
...
      "BlacklistRequest": {
        "properties": {
          "company": {
            "description": "Nom de l'entreprise à exclure",
            "minLength": 1,
            "title": "Company",
            "type": "string"
          }
        },
        "required": [
          "company"
        ],
        "title": "BlacklistRequest",
        "type": "object"
      },
      "BlacklistResponse": {
        "properties": {
          "excluded_companies": {
            "items": {
              "type": "string"
            },
            "title": "Excluded Companies",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retro_updated_jobs": {
            "default": 0,
            "description": "Offres existantes de cette entreprise passées en BLACKLISTED",
            "title": "Retro Updated Jobs",
            "type": "integer"
          }
        },
        "required": [
          "message",
          "excluded_companies"
        ],
        "title": "BlacklistResponse",
        "type": "object"
      },
      "CacheStats": {
        "additionalProperties": true,
        "properties": {
          "hit_ratio_percent": {
            "default": 0.0,
            "title": "Hit Ratio Percent",
            "type": "number"
          },
          "hits": {
            "default": 0,
            "title": "Hits",
            "type": "integer"
          },
          "misses": {
            "default": 0,
            "title": "Misses",
            "type": "integer"
          },
          "size": {
            "default": 0,
            "title": "Size",
            "type": "integer"
          }
        },
        "title": "CacheStats",
        "type": "object"
      },
      "CandidatePreferences": {
        "properties": {
          "ai_custom_instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "title": "Ai Custom Instructions"
          },
          "ai_model": {
            "default": "gemini-2.5-flash",
            "title": "Ai Model",
            "type": "string"
          },
          "ai_temperature": {
            "default": 0.2,
            "title": "Ai Temperature",
            "type": "number"
          },
          "auto_prepare_documents": {
            "default": false,
            "title": "Auto Prepare Documents",
            "type": "boolean"
          },
          "auto_sync_notion": {
            "default": true,
            "title": "Auto Sync Notion",
            "type": "boolean"
          },
          "contract_types": {
            "items": {
              "type": "string"
            },
            "title": "Contract Types",
            "type": "array"
          },
          "daily_collection_limit": {
            "default": 5,
            "title": "Daily Collection Limit",
            "type": "integer"
          },
          "default_search_duration": {
            "default": "24h",
            "title": "Default Search Duration",
            "type": "string"
          },
          "excluded_companies": {
            "items": {
              "type": "string"
            },
            "title": "Excluded Companies",
            "type": "array"
          },
          "excluded_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Excluded Keywords",
            "type": "array"
          },
          "filter_esn": {
            "default": false,
            "title": "Filter Esn",
            "type": "boolean"
          },
          "locations": {
            "items": {
              "type": "string"
            },
            "title": "Locations",
            "type": "array"
          },
          "match_threshold_priority": {
            "default": 85,
            "title": "Match Threshold Priority",
            "type": "integer"
          },
          "match_threshold_recommended": {
            "default": 75,
            "title": "Match Threshold Recommended",
            "type": "integer"
          },
          "match_threshold_review": {
            "default": 60,
            "title": "Match Threshold Review",
            "type": "integer"
          },
          "minimum_salary": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minimum Salary"
          },
          "remote": {
            "default": true,
            "title": "Remote",
            "type": "boolean"
          },
          "sectors": {
            "items": {
              "type": "string"
            },
            "title": "Sectors",
            "type": "array"
          },
          "target_titles": {
            "items": {
              "type": "string"
            },
            "title": "Target Titles",
            "type": "array"
          }
        },
        "title": "CandidatePreferences",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProviderInfo": {
        "properties": {
          "default_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Model"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "is_configured": {
            "title": "Is Configured",
            "type": "boolean"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "title": "Models",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "stats": {
            "$ref": "#/components/schemas/ProviderStats"
          }
        },
        "required": [
          "name",
          "display_name",
          "is_configured",
          "stats"
        ],
        "title": "ProviderInfo",
        "type": "object"
      },
      "ProviderStats": {
        "additionalProperties": true,
        "properties": {
          "errors_count": {
            "default": 0,
            "title": "Errors Count",
            "type": "integer"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "requests_count": {
            "default": 0,
            "title": "Requests Count",
            "type": "integer"
          },
          "success_count": {
            "default": 0,
            "title": "Success Count",
            "type": "integer"
          },
          "total_latency_seconds": {
            "default": 0.0,
            "title": "Total Latency Seconds",
            "type": "number"
          },
          "total_tokens": {
            "default": 0,
            "title": "Total Tokens",
            "type": "integer"
          }
        },
        "required": [
          "provider"
        ],
        "title": "ProviderStats",
        "type": "object"
      },
      "ProviderTestRequest": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "provider": {
            "default": "gemini",
            "title": "Provider",
            "type": "string"
          }
        },
        "title": "ProviderTestRequest",
        "type": "object"
      },
      "ProviderTestResponse": {
        "additionalProperties": true,
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "response_time_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Time Ms"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "ProviderTestResponse",
        "type": "object"
      },
...
      "RouterOverview": {
        "properties": {
          "cache": {
            "$ref": "#/components/schemas/CacheStats"
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderInfo"
            },
            "title": "Providers",
            "type": "array"
          },
          "routing_config": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "description": "Ordre de repli des providers par type de tâche",
            "title": "Routing Config",
            "type": "object"
          }
        },
        "required": [
          "providers",
          "routing_config",
          "cache"
        ],
        "title": "RouterOverview",
        "type": "object"
      },
...
      "SettingsResponse": {
        "properties": {
          "available_models": {
            "items": {
              "$ref": "#/components/schemas/AIModel"
            },
            "title": "Available Models",
            "type": "array"
          },
          "preferences": {
            "$ref": "#/components/schemas/CandidatePreferences"
          },
          "router_overview": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RouterOverview"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "preferences",
          "available_models"
        ],
        "title": "SettingsResponse",
        "type": "object"
      },
...
      "UpdateSettingsRequest": {
        "properties": {
          "preferences": {
            "$ref": "#/components/schemas/CandidatePreferences"
          }
        },
        "required": [
          "preferences"
        ],
        "title": "UpdateSettingsRequest",
        "type": "object"
      },
      "UpdateSettingsResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "preferences": {
            "$ref": "#/components/schemas/CandidatePreferences"
          }
        },
        "required": [
          "message",
          "preferences"
        ],
        "title": "UpdateSettingsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Backend API moderne pour l'agent IA de recherche d'emploi et de génération de candidatures sur mesure.",
    "title": "Trema Job Search API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "description": "Récupère les préférences et paramètres actuels.",
        "operationId": "get_settings_api_v1_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Settings",
        "tags": [
          "Settings"
        ]
      },
      "put": {
        "description": "Met à jour les paramètres dans le profil candidat actif.",
        "operationId": "update_settings_api_v1_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Settings",
        "tags": [
          "Settings"
        ]
      }
    },
    "/api/v1/settings/blacklist": {
      "delete": {
        "description": "Retire une entreprise de la liste noire (les offres déjà marquées BLACKLISTED ne sont pas modifiées).",
        "operationId": "remove_from_blacklist_api_v1_settings_blacklist_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlacklistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlacklistResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove From Blacklist",
        "tags": [
          "Settings"
        ]
      },
      "post": {
        "description": "Ajoute une entreprise à la liste noire et passe ses offres existantes en BLACKLISTED.",
        "operationId": "add_to_blacklist_api_v1_settings_blacklist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlacklistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlacklistResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add To Blacklist",
        "tags": [
          "Settings"
        ]
      }
    },
...
    "/api/v1/settings/router/test": {
      "post": {
        "description": "Teste la connectivité d'un provider LLM (Gemini, Groq, Mistral, OpenRouter) avec un modèle optionnel.",
        "operationId": "test_router_provider_api_v1_settings_router_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderTestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Router Provider",
        "tags": [
          "Settings"
        ]
      }
    },
...
  }
}